Update calculations to coyote and blower dynamics
Refactor the model by introducing better granularity to coyote's deceleration phases to make the prediction more accurate. Now, the coyote has two phases for deceleration instead of one. Adjust the formulae used for calculating catchup time, data for acceleration and deceleration accordingly. Changes are made to "get_acceleration_coyote" and "get_acceleration_blower" to accommodate the new dynamics. This will improve the accuracy of the simulation and data representation.